pyproject.toml
[tool.poetry]
name = "zstod"
version = "0.1.0"
description = ""
authors = ["Adib Mosharrof <adib.mosharrof@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.3.0"
pathos = "^0.3.0"
pyhumps = "^3.8.0"
hydra-core = "^1.3.2"
matplotlib = "^3.7.1"
fuzzywuzzy = "^0.18.0"
evaluate = "^0.4.0"
dotmap = "^1.3.30"
dataclasses-json = "^0.5.7"
dataclass-csv = "^1.4.0"
transformers = {git = "https://github.com/huggingface/transformers.git", rev = "main"}
wandb = "^0.14.0"
sentence-transformers = "^2.2.2"
seaborn = "^0.12.2"
numerize = "^0.12"
torchmetrics = "^1.2.1"
accelerate = "^0.25.0"
loralib = "^0.1.2"
absl-py = "^1.4.0"
rouge-score = "^0.1.2"
sentencepiece = "^0.1.97"
peft ="^0.6.2" 
omegaconf = "^2.3.0"
sgd-dstc8-data-model = "^0.1.9"
trl = "^0.7.4"
bert-score = "^0.3.13"
hurry = "^1.1"
hurry-filesize = "^0.9"
bitsandbytes = "^0.41.2.post2"
scipy = "^1.11.4"
torch = "2.0.1"
deepspeed = "^0.12.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
